+++ src/main.jsx
@@ -2,9 +2,26 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App.jsx';
+import { Provider } from 'react-redux';
+import store from './redux/store';
+import { PersistGate } from 'redux-persist/integration/react';
+import { persistStore } from 'redux-persist';
+import { CookiesProvider } from 'react-cookie';
+import { QueryClientProvider, QueryClient } from '@tanstack/react-query';
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 
+const queryClient = new QueryClient();
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <App />
+    <QueryClientProvider client={queryClient}>
+      <ReactQueryDevtools initialIsOpen={true} />
+      <Provider store={store}>
+        <PersistGate loading={null} persistor={persistStore(store)}>
+          <CookiesProvider>
+            <App />
+          </CookiesProvider>
+        </PersistGate>
+      </Provider>
+    </QueryClientProvider>
   </StrictMode>,
 );
